Rename plainmap -> maps.
... now that all the source code is in that.
